requirements.txt
bokeh==2.0.1
click==7.1.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.2
packaging==20.3
pandas==1.0.3
Pillow==7.1.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
tornado==6.0.4
typing-extensions==3.7.4.2
Werkzeug==1.0.1
gunicorn==20.0.4
requests==2.23.0
Flask-MQTT==1.1.0
Flask-Cors==3.0.8
psycopg2==2.8.5
Flask-SQLAlchemy==2.4.1
PyMySQL==0.9.3
